projectwork/project-rag-istat/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pymupdf
python-dotenv
networkx

# optional 
fastapi
uvicorn[standard]
ragas
# telegram
python-telegram-bot
